+++ CHANGELOG.md
@@ -2,6 +2,48 @@
 
 <!--- next entry here -->
 
+## 1.4.0
+2020-01-29
+
+### Features
+
+- adding commandline arg support (2632ed72f7127061522c192b78b360a62baeec38)
+- Added more args (a461ff07b1bd00169d68bb56d85e0f53e90033d1)
+- finished adding argv support for tap_tenant (a7ce6961d0accb56c7987ccfee53b5e6128a344c)
+- removed dead code (8a842369d0f68c9965c804e51d1c7abf0f1b1e9a)
+- removed dead code (3a80c4bc4c6707ee75f6bd5c64eaa6d9658d97e3)
+- removed dead code (bc6ccf3502ae25fe6542c0fb0dcc8cc269b098c3)
+- removed dead code (086755c319792f746844efaadea2830a9396f8bd)
+- add mac (b21748cf87761b2f281e7f0ea466c61f03d8c6b0)
+
+## 1.3.4
+2020-01-29
+
+### Fixes
+
+- checking for mod 4 alignment in descosck mem pools (fcd8fbe492c7188436289e89a221ad2c9a20a93a)
+- check for mod 4 alignment (07ffc7e4fab24c9c9c95b2e465d4ca9db6cac3c6)
+- renamed function (5383bd06b7e5d6a5e19a6602c9f6d69b34aa9735)
+- add counter for non 4b aligned buffers (57502018948b75cfff95fa91a22011c3f4446c3c)
+- add stat counter for non 4byte alinged buffers (e7f9cad535526ad5526d65532d5668cf0b4fc98e)
+- removed commented out code, renamed var (fbc5a01abb873773bab63a9ca1923416785f5a3e)
+
+## 1.3.3
+2020-01-23
+
+### Fixes
+
+- initialize send descriptor flags (544ac6796f246e3e08c492275e1de36e360eb789)
+- initialize send descriptor flags to 0 (b73078021899a06ef6f75c15e5f69ddb2e4de190)
+- initialize send descriptor flags to 0 (3e001fd7c91cda9ff4671c83e123c577a1c3fed7)
+
+## 1.3.2
+2020-01-09
+
+### Fixes
+
+- Correct laden descriptor size, add static assert (662bf8e6a57aa7996e3cdec4afd20dfea8b6b262)
+
 ## 1.3.1
 2019-12-06